tests/wpt/web-platform-tests/bluetooth/service/getCharacteristics/characteristics-not-found.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
const test_desc = 'Request for absent characteristics. Reject with ' +
    'NotFoundError.';
const expected = new DOMException('No Characteristics found in service.',
    'NotFoundError');
let service;

bluetooth_test(() => getEmptyHealthThermometerService()
    .then(_ => ({service} = _))
    .then(() => assert_promise_rejects_with_message(
      service.getCharacteristics(), expected)),
    test_desc);
</script>
